Changed db_update script to perform the table convertion to UTF-8 first, cancelling the need to work with latin1/UTF-8 connection. Now the connection can always been purely UTF-8.
Before we converted the data to UTF-8 but it was stored in a latin1 table, meaning the string length was calculated wrongly for multibyte characters and hence the max field size could be reached when it shouldn't be.

Hopefully this fixes the issues, but it would be nice if anyone with a 1.2 forum with a large amount of non-ascii contents could give it a test.

http://fluxbb.org/forums/topic/3570/update-proccess-fixed-length-in-fields/