Adding support for searching for CJK (Chinese, Japanese, Korean) words.
Because CJK words are not split by spaces the current indexing function will not index them properly. We could assume that 1 character = 1 word and split by that, but it isn't always correct, and results in a huge (~10x?) search index. This method detects CJK words when a search is performed, and switches to a much simplier (but much less efficient) method of searching each post individually for the requested word.

Other than checking if a word is a CJK word or normal word before performing a search this should have no effect on non-CJK boards.