Allow nested code tags.
Well, almost. Testing the `split_text_test` in a stand-alone script
works, but not in FluxBB, because the BBcode validator in the parser
gets in the way saying there's a missing opening tag.

The following:

    Test test

    [code]This is [code]s[code]om[/code]e[/code] code.[/code]

    [code]This [code]is[/code] some more [code]s[code]om[/code]e[/code]
    code.[/code]

Should result in:

    Test test

    This is [code]s[code]om[/code]e[/code] code.

    This [code]is[/code] some more [code]s[code]om[/code]e[/code] code.

But the parser returns this error: [/code] was found without a matching [code]